microservices/bg-removal-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aio-pika==9.3.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2